Website_GPS/requirements.txt
alabaster==0.7.8
apturl==0.5.2
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
brial==1.0.2
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cvxopt==1.2.3
cycler==0.10.0
cypari2==2.1.1
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
fpylll==0.5.1.dev0
future==0.18.2
gcloud==0.18.3
GDAL==3.0.4
gmpy2==2.1.0b3
googleapis-common-protos==1.56.3
gpg===1.13.1-unknown
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jdcal==1.0
jedi==0.15.2
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jwcrypto==1.3.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
ovs==2.13.3
OWSLib==0.19.1
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parameterized==0.7.0
paramiko==2.6.0
parso==0.5.2
path.py==12.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
pkgconfig==1.5.1
plotly==4.4.1
pluggy==0.13.0
pplpy==0.8.4
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.4
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.15.0
pycups==1.9.73
pydot==1.4.1
Pygments==2.3.1
PyGObject==3.36.0
pygpu==0.7.6
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymol==2.3.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
Pyrebase4==4.5.0
pyRFC3339==1.1
pyrsistent==0.15.5
PyStemmer==1.3.0
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jwt==3.3.2
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
roman==2.0.0
rpy2==3.2.6
rsa==4.8
sage==9.0
sagenb-export==3.2
scikit-learn==0.22.2.post1
scipy==1.3.3
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-scm==3.4.3
simplegeneric==0.8.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==1.8.5
sqlparse==0.2.4
ssh-import-id==5.10
sshuttle==0.78.5
sympy==1.5.1
systemd-python==234
tables==3.6.1
terminado==0.8.2
testpath==0.4.4
Theano==1.0.4
toml==0.10.0
tornado==5.1.1
traitlets==4.3.3
Twisted==18.9.0
tzlocal==2.1b1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcversioner==2.16.0.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==2.0.0
wrapt==1.14.1
wxPython==4.0.7
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
yarg==0.1.9
zipp==1.0.0
zope.interface==4.7.1
